Factor to components (#2588)
* very rough layout update

* Remove filtering into showcase columns

* added DashboardPublic + added route to use DashboardPublic on no signedIn

* finished basic format for dashboardPublic.js

* added fetch for showcase.json
rough sketch

* Show all workspaces from json

* rough side bar

* improvements to card rendering and sizing

* get count of labels

* case-insensitive tag filtering

* added search (still needs styling)

* added showcase sort

* showcase multi-tags

* reworked the layout to make sure the title, search, and sort-by can all be baseline aligned

* `function` to => function

* updated categories and tags
fixed a bug with multi-tag filter

* added collapsing sections to showcase.js

* pills

* suppress labels with no matching workspaces

* lots of style tweaks to get it much closer to the original designs

* rework filtering to update counts as filters are applied

* simplified tag toggling code

* fixed sidebar initial open state
fixed react warning

* added special handling of Getting Started section

* minor case tweaks

* moved `sidebarSections` to the top of the file to make it easier to find and maintain

* make text search case-insensitive

* PR feedback

* additional PR feedback from Mohs (and prior changes to SignInButton.js that I somehow missed committing before)

* PR feedback from Isaac (and a bugfix on the 'clear' button)

* another change inspired by Isaac's feedback

* memoize groupByFeaturedTags

* Factored everything but showcase component

* sort of working

* more refactors, run lint

* clean up unused imports

* fix tab switching, clean up imports

* lint

* incorporate pr feedback

* lint

* remove unused import

* fix bug with StateHistory
